safari-600.7-branch/LayoutTests/platform/ios-sim/fast/forms/datetime/datetime-stepup-stepdown-expected.txt

Check stepUp() and stepDown() bahevior for type=datetime.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Invalid value
PASS stepUp("", null, null) threw exception Error: InvalidStateError: DOM Exception 11.
PASS stepDown("", null, null) threw exception Error: InvalidStateError: DOM Exception 11.
Non-number arguments
FAIL stepUp("2010-02-10T20:13Z", null, null, "0") should be 2010-02-10T20:13Z. Threw exception Error: InvalidStateError: DOM Exception 11
FAIL stepDown("2010-02-10T20:13Z", null, null, "0") should be 2010-02-10T20:13Z. Threw exception Error: InvalidStateError: DOM Exception 11
FAIL stepUp("2010-02-10T20:13Z", null, null, "foo") should be 2010-02-10T20:13Z. Threw exception Error: InvalidStateError: DOM Exception 11
FAIL stepDown("2010-02-10T20:13Z", null, null, "foo") should be 2010-02-10T20:13Z. Threw exception Error: InvalidStateError: DOM Exception 11
FAIL stepUp("2010-02-10T20:13Z", null, null, null) should be 2010-02-10T20:13Z. Threw exception Error: InvalidStateError: DOM Exception 11
FAIL stepDown("2010-02-10T20:13Z", null, null, null) should be 2010-02-10T20:13Z. Threw exception Error: InvalidStateError: DOM Exception 11
Normal cases
FAIL stepUp("2010-02-10T20:13Z", null, null) should be 2010-02-10T20:14Z. Threw exception Error: InvalidStateError: DOM Exception 11
FAIL stepDown("2010-02-10T20:13Z", null, null) should be 2010-02-10T20:12Z. Threw exception Error: InvalidStateError: DOM Exception 11
FAIL stepUp("2010-02-10T20:13Z", null, null, 10) should be 2010-02-10T20:23Z. Threw exception Error: InvalidStateError: DOM Exception 11
FAIL stepDown("2010-02-10T20:13Z", null, null, 11) should be 2010-02-10T20:02Z. Threw exception Error: InvalidStateError: DOM Exception 11
FAIL stepUp("1970-01-01T20:13Z", "4", null, 2) should be 1970-01-01T20:13:08Z. Threw exception Error: InvalidStateError: DOM Exception 11
FAIL stepDown("1970-01-01T20:13Z", "4", null, 3) should be 1970-01-01T20:12:48Z. Threw exception Error: InvalidStateError: DOM Exception 11
Step=any
PASS stepUp("2010-02-10T20:13Z", "any", null) threw exception Error: InvalidStateError: DOM Exception 11.
PASS stepDown("2010-02-10T20:13Z", "any", null) threw exception Error: InvalidStateError: DOM Exception 11.
Overflow/underflow
PASS stepUp("2010-02-10T20:13Z", "3.40282346e+38", null) threw exception Error: InvalidStateError: DOM Exception 11.
PASS stepDown("2010-02-10T20:13Z", "3.40282346e+38", null) threw exception Error: InvalidStateError: DOM Exception 11.
PASS stepUp("2010-02-10T20:13Z", "1", "2010-02-10T20:13Z") threw exception Error: InvalidStateError: DOM Exception 11.
PASS stepDown("2010-02-10T20:13Z", "1", "2010-02-10T20:13Z") threw exception Error: InvalidStateError: DOM Exception 11.
PASS successfullyParsed is true

TEST COMPLETE
